@@ -12,20 +12,24 @@
   "dependencies": {
     "css-reset": "git+https://github.com/60frames/css-reset.git",
     "lodash.throttle": "^3.0.4",
-    "react": "^15.2.0",
-    "react-dom": "^15.2.0"
+    "prop-types": "^15.6.0",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0"
   },
   "devDependencies": {
-    "babel-core": "^5.8.25",
-    "babel-loader": "^5.3.2",
-    "babel-runtime": "^6.3.13",
-    "coding-standards": "git+ssh://git@github.com/60frames/coding-standards.git#1.0.5",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^6.4.1",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-react": "^6.24.1",
     "css-loader": "^0.21.0",
     "eslint": "^1.8.0",
     "eslint-plugin-react": "^3.6.3",
     "file-loader": "^0.8.4",
     "react-hot-loader": "^1.3.0",
     "style-loader": "^0.13.0",
+    "url-loader": "^0.6.2",
     "webpack": "^1.12.2",
     "webpack-dev-server": "^1.12.1"
   }
