@@ -2,32 +2,30 @@
 /*
  * Generated by the 'fontello-react' Grunt task.
  */
+import PropTypes from 'prop-types';
+
 import React from 'react';
 
-var Icon = React.createClass({
-
-    propTypes: {
-        name: React.PropTypes.oneOf([
+class Icon extends React.Component {
+    static propTypes = {
+        name: PropTypes.oneOf([
             'github-circled-alt2'
         ])
-    },
+    };
 
     /**
      * Default the icon to the first one just to show something
      * @return {Object} The default props
      */
-    getDefaultProps() {
-        return {
-            name: 'github-circled-alt2'
-        };
-    },
+    static defaultProps = {
+        name: 'github-circled-alt2'
+    };
 
     render() {
         return (
             <span className={'icon icon--' + this.props.name}></span>
         );
     }
-
-});
+}
 
 export default Icon;
