@@ -2,12 +2,13 @@
 /*
  * Generated by the 'fontello-react' Grunt task.
  */
+import PropTypes from 'prop-types';
+
 import React from 'react';
 
-var Icon = React.createClass({
-
-    propTypes: {
-        name: React.PropTypes.oneOf([
+class Icon extends React.Component {
+    static propTypes = {
+        name: PropTypes.oneOf([
             'play-1',
 			'volume-off',
 			'volume-down',
@@ -16,24 +17,21 @@ var Icon = React.createClass({
 			'resize-small',
 			'pause-1'
         ])
-    },
+    };
 
     /**
      * Default the icon to the first one just to show something
      * @return {Object} The default props
      */
-    getDefaultProps() {
-        return {
-            name: 'play-1'
-        };
-    },
+    static defaultProps = {
+        name: 'play-1'
+    };
 
     render() {
         return (
             <span className={'video-icon video-icon--' + this.props.name}></span>
         );
     }
-
-});
+}
 
 export default Icon;
